pyproject.toml
[project]
name = "ascii-api"
version = "0.1.0"
description = "FastAPI REST API service for converting images to ASCII art"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.12",
    "python-magic>=0.4.27",
    "fastapi-limiter>=0.1.6",
    "python-jose[cryptography]>=3.3.0",
    "structlog>=25.0.0",
    "prometheus-client>=0.21.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "aiofiles>=24.1.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov==7.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
    "N",   # pep8-naming
    "ANN", # flake8-annotations
]
ignore = [
    "S101", # use of assert (needed in tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
